refactor(*): change directory structure & rename injection token
