+++ docs/interfaces/event.md
@@ -11,4 +11,4 @@
 
 ## Source
 
-[lib/models/models.ts](https://github.com/ddubrava/angular8-yandex-maps/blob/master/projects/angular8-yandex-maps/src/lib/models/models.ts)
+[lib/interfaces/event.ts](https://github.com/ddubrava/angular8-yandex-maps/blob/master/projects/angular8-yandex-maps/src/lib/interfaces/event.ts)
