+++ docs/quickstart.md
@@ -47,13 +47,13 @@ export class AppModule {}
 Injection token
 
 ```ts
-import { AngularYandexMapsModule, YA_MAP_CONFIG } from 'angular8-yandex-maps';
+import { AngularYandexMapsModule, YA_CONFIG } from 'angular8-yandex-maps';
 
 @NgModule({
   imports: [AngularYandexMapsModule],
   providers: [
     {
-      provide: YA_MAP_CONFIG,
+      provide: YA_CONFIG,
       useValue: {
         apikey: 'API_KEY',
         lang: 'en_US',
