+++ projects/angular8-yandex-maps/src/lib/constants/constant.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+import { YaConfig } from '../interfaces/config';
+
+export const YA_CONFIG = new InjectionToken<YaConfig>('YA_CONFIG');
