+++ projects/angular8-yandex-maps/src/lib/directives/ya-geoobject/ya-geoobject.directive.ts
@@ -10,7 +10,7 @@ import {
   SimpleChanges,
 } from '@angular/core';
 import { generateRandomId } from '../../utils/generateRandomId';
-import { IEvent, ILoadEvent } from '../../models/models';
+import { IEvent, ILoadEvent } from '../../interfaces/event';
 import { removeLeadingSpaces } from '../../utils/removeLeadingSpaces';
 
 /**
