+++ projects/angular8-yandex-maps/src/lib/interfaces/config.ts
@@ -1,5 +1,3 @@
-import { InjectionToken } from '@angular/core';
-
 /**
  * API loading parameters
  * @see {@link https://yandex.ru/dev/maps/jsapi/doc/2.1/dg/concepts/load.html/#load__param} for further information
@@ -34,36 +32,3 @@ export interface YaConfig {
    */
   version?: string;
 }
-
-export interface ILoadEvent {
-  /**
-   * Instance of created entity
-   */
-  instance?: any;
-  /**
-   * API global object
-   */
-  ymaps: typeof ymaps;
-}
-
-export interface IEvent {
-  /**
-   * Instance of created entity
-   */
-  instance: any;
-  /**
-   * API global object
-   */
-  ymaps: typeof ymaps;
-  /**
-   * String event type, event.originalEvent.type
-   */
-  type: string | undefined;
-  /**
-   * Provides methods for accessing the originalObject object's fields and methods, with the possibility for redefining them.
-   * @see {@link https://yandex.ru/dev/maps/jsapi/doc/2.1/ref/reference/Event.html/}
-   */
-  event: any;
-}
-
-export const YA_MAP_CONFIG = new InjectionToken<YaConfig>('YA_MAP_CONFIG');
